% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_kraken_data.R
\name{down_multi_ass}
\alias{down_multi_ass}
\title{download multiple assets}
\usage{
down_multi_ass(ass_l, folder_root, to_long = TRUE)
}
\description{
call down_tick_data for multiple assets and create the long
file if flagged
}
