% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_kraken_data.R
\name{down_tick_data}
\alias{down_tick_data}
\title{down_tick_data}
\usage{
down_tick_data(asset_in, folder_root)
}
\description{
access functio to take in the asset name, get the since and then
download the data
}
