% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_kraken_data.R
\name{get_historical_trades}
\alias{get_historical_trades}
\title{script to download kraken tick data}
\usage{
get_historical_trades(pair_in, folder_in, curr_since)
}
\description{
uses the krakenR package to download all historical tick
data for the specified pair
}
