% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_proc.R
\name{NOT_USED_error_proc_fun}
\alias{NOT_USED_error_proc_fun}
\title{process the errors from check_long_df}
\usage{
NOT_USED_error_proc_fun(curr_he_in, kraken_root_dir_in)
}
\description{
Convert data downloaded from kraken to long dataframe
}
