% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ohlc.R
\name{to_ohlc}
\alias{to_ohlc}
\title{to ohlc}
\usage{
to_ohlc(
  in_long,
  dp_dir_in,
  periods = c("1 minute", "30 minutes", "1 hour", "4 hours", "12 hours", "daily")
)
}
\description{
convert the long_df from to_long to ohlc
}
