% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_ohlc.R
\name{down_ohlc}
\alias{down_ohlc}
\title{download ohlc for an asset pair}
\usage{
down_ohlc(asset_pair, min_date, max_date, interval)
}
\description{
download ohlc between specified dates, of a specified period for
the specified asset pair
}
