% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_kraken_data.R
\name{get_historical_trades}
\alias{get_historical_trades}
\title{script to download kraken tick data}
\usage{
get_historical_trades(pair_in, folder_in, curr_since, max_retries = 3)
}
\description{
uses the krakenR package to download all historical tick
data for the specified pair
pair in e.g.: ETHEUR
file_in e.g.: path/file.csv
}
