% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ohlc.R
\name{to_ohlc}
\alias{to_ohlc}
\title{to ohlc}
\usage{
to_ohlc(
  long_df,
  dp_dir_in,
  periods = c("1 seconds", "30 seconds", "1 minutes", "30 minutes", "1 hours",
    "4 hours", "12 hours", "24 hours")
)
}
\description{
convert the long_df from to_long to ohlc
}
